requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiosqlite==0.13.0
altgraph==0.16.1
async-timeout==3.0.1
attrs==20.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cloudscraper==1.2.46
colorama==0.4.1
colored==1.4.0
cryptography==2.8
Cython==0.29.15
dnslib==0.9.10
dnspython==1.16.0
elasticsearch==7.1.0
et-xmlfile==1.0.1
Flask==1.1.1
FOFA==1.0.1
geoip2==4.0.2
gevent
google==2.0.3
greenlet==0.4.15
idna==2.8
IPy==1.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kazoo==2.6.1
ldap3==2.6.1
lxml==4.5.1
macholib==1.11
MarkupSafe==1.1.1
maxminddb==2.0.2
multidict==4.7.6
netaddr==0.7.19
openpyxl==2.6.4
paramiko==2.7.1
plotly==4.7.1
pyasn1==0.4.8
pycares==3.1.1
pycparser==2.19
pycryptodome==3.9.2
PyInstaller==3.5
pymongo==3.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
PySocks==1.7.1
PythonDNS==0.1
redis==3.3.11
requests==2.24.0
requests-file==1.4.3
requests-toolbelt==0.9.1
retrying==1.3.3
scapy==2.4.3
shodan==1.7.5
six==1.13.0
soupsieve==1.9.5
termcolor==1.1.0
texttable==1.6.2
tldextract==2.2.2
urllib3==1.25.8
uvloop==0.14.0
Werkzeug==1.0.0
XlsxWriter==1.2.5
xlwt==1.3.0
yarl==1.5.1
PyYAML==5.3.1
tqdm==4.51.0
